+++ NEWS.md
@@ -1,5 +1,9 @@
 # croquet (development version)
 
+* Added functions to select all upper case (`all_uppercase()`), lower case (`all_lowercase()`), and columns with label attributes (`all_labelled()`).
+
+* Re-exporting the `dplyr::select()` function.
+
 # croquet 0.1.0
 
 * Initial release.